Make call to zip UNB files a separate step
